Module Zmq_lwt__Deferred.Fd

type t

val create : Unix.file_descr -> t

val wait_readable : t -> unit Deferred.t

Wait for the fd to become readable. It is important that all waiters on the socket are woken up, and not just one, as the fd is used as a broadcast mechanism.

The Fd must _NOT_ be closed, as its owned by Zmq.

val release : t -> unit Deferred.t